+++ lib/shrine/plugins/default_url_options.rb
@@ -11,7 +11,7 @@ class Shrine
     # block, which will receive the UploadedFile object along with any options
     # that were passed to `UploadedFile#url`.
     #
-    #     plugin :default_url_options, store: ->(io, **options) do
+    #     plugin :default_url_options, store: -> (io, **options) do
     #       { response_content_disposition: ContentDisposition.attachment(io.original_filename) }
     #     end
     #
@@ -20,7 +20,8 @@ class Shrine
     # default options.
     module DefaultUrlOptions
       def self.configure(uploader, options = {})
-        uploader.opts[:default_url_options] = (uploader.opts[:default_url_options] || {}).merge(options)
+        uploader.opts[:default_url_options] ||= {}
+        uploader.opts[:default_url_options].merge!(options)
       end
 
       module FileMethods